[API] Rename account recovery route
