+++ api/lib/routes.js
@@ -1,5 +1,5 @@
 module.exports = [
-  require('./application/account-recoveries'),
+  require('./application/account-recovery'),
   require('./application/answers'),
   require('./application/assessment-results'),
   require('./application/assessments'),
